resolves #567 Elements in Duplicated Submodels show the correct submodel names
When duplicating a submodel, the elements were initialized while the submodel still had exactly the same name as the original. Because of this, in the ModelBrowser there were always 2 times the same submodel behind the element name in the beginning and not once the old submodel name and once the new one.