refs #494: Base for loading model templates
Now all files in the fixed tool template directory are searched and it is determined whether it is a parsable 3LGM2 file and if yes, to which metamodel this file fits. This allows TemplateLibraryManager to determine which model files it would need to provide as templates for a loaded model.