refs #580 Context Sensitive Select All
Select All with focus in graph: only visible Elements on the single layer or if multi layer view is active on all layers will be selected.Select all in ModelBrowser: all elements shown in the browser are selected. Select all in TemplateBrowser: all nodes in the template tree are selected. Select All in Consistency