refs #584 Prevent Exception on Copy from M to S
This commit also enables basic functionality to really copy elements between different metamodels. Especially nodes and egdes that have an 1:1 equivalent in the other metamdel can be copied. But for now there are only a few node and edge classes mapped from 3LGM2-M to 3LGM2-S, but nothing in the other direction.